Добавил интерфейс TaskManager и класс с реализацией методов интерфейса InMemoryTaskManager (Принцип ООП)
