+++ src/tracker/Main.java
@@ -1,11 +1,12 @@
 package tracker;
 
+import tracker.controllers.InMemoryTaskManager;
 import tracker.model.*;
 import tracker.controllers.TaskManager;
 
 public class Main {
     public static void main(String[] args) {
-        TaskManager manager = new TaskManager();
+        TaskManager manager = new InMemoryTaskManager();
 
         // Создание 2 задач
         Task task1 = new Task("Сходить в магазин", "Купить продукты", Status.NEW);
